v2_professional/backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
faiss-cpu
pypdf
python-docx
sentence-transformers
torch
python-dotenv
